requirements.txt
python-telegram-bot>=21.7,<22.0
python-dotenv>=1.0.1,<2.0.0
pywinauto>=0.6.8,<0.7.0
pywin32>=308
pyautogui>=0.9.54,<1.0.0
pyperclip>=1.9.0,<2.0.0
Pillow>=11.3.0,<13.0.0
pytesseract>=0.3.13,<0.4.0
rapidocr-onnxruntime>=1.2.3,<1.3.0
numpy>=1.26.4,<3.0.0
